requirements.txt
fastapi
uvicorn
python-dotenv
crewai
crewai_tools
openai
google-generativeai
